+++ .gitignore
@@ -1,2 +1,2 @@
 node_modules
-".DS_Store".DS_Store
+".DS_Store"
