+++ js/main.jsx
@@ -1,7 +1,7 @@
 var React = require('react');
 var ReactDOM = require('react-dom');
 var Backbone = require('backbone');
-var List = require('../components/List.jsx');
+var List = require('/components/list.jsx');
 
 
-ReactDOM.render(<List />, document.getElementById('ListCom'));
+ReactDOM.render(<Input />, document.getElementById('ListCom'));
